Make round functions use int64s
We're getting some integer overflows at big layer heights, causing the slice to become empty because it thinks it needs a negative amount of layers.
The root of the matter is the multiplication done to determine the sign (dividend * divisor) which causes the integer overflow. This is not a problem for some of these functions, but to be consistent I changed all of them.

It should also improve performance a bit by preventing the need for casts to smaller integer numbers. These functions are almost exclusively called with coord_t and size_t, which are both 64-bit.

Fixes a problem reported on Reddit: https://www.reddit.com/r/Cura/comments/rbdu07/maximum_layer_height_cura_for_wire_arc_additive/

+++ src/skin.cpp
@@ -698,7 +698,7 @@ void SkinInfillAreaComputation::combineInfillLayers(SliceMeshStorage& mesh)
     }
 
     const coord_t layer_height = mesh.settings.get<coord_t>("layer_height");
-    const size_t amount = std::max(1U, round_divide(mesh.settings.get<coord_t>("infill_sparse_thickness"), std::max(layer_height, coord_t(1)))); //How many infill layers to combine to obtain the requested sparse thickness.
+    const size_t amount = std::max(uint64_t(1), round_divide(mesh.settings.get<coord_t>("infill_sparse_thickness"), std::max(layer_height, coord_t(1)))); //How many infill layers to combine to obtain the requested sparse thickness.
     if(amount <= 1) //If we must combine 1 layer, nothing needs to be combined. Combining 0 layers is invalid.
     {
         return;
